requirements/base.txt
django-environ==0.4.5
djangorestframework >= 3.12, < 3.13
djangorestframework-simplejwt >= 4.7, < 4.8
django-cors-headers >= 3.7.0, < 4
drf-yasg >= 1.20.0, < 2
